api-gateway/requirements.txt
# API Gateway Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP client
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis for caching and rate limiting
redis==5.0.1

# Settings management
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Logging
python-json-logger==2.0.7

# Prometheus metrics
# prometheus-client==0.19.0
# Monitoring
prometheus-fastapi-instrumentator

# OpenTelemetry Tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger-thrift
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
structlog

# Environment variables
python-dotenv==1.0.0